// shared-schemas/src/types.d.ts
export interface Point2D {
    x_mm: number;
    y_mm: number;
}
// ...
export interface CeilingConstraint {
    id?: string;
    room_id?: string;
    wall_id: string;
    wall_start: Point2D;
    wall_end: Point2D;
    kniestock_height_mm: number;
    slope_angle_deg: number;
    depth_into_room_mm: number;
}